Data Analyst Copilot/requirements.txt
openai
streamlit
ydata-profiling
streamlit-chat
python-dotenv
matplotlib
seaborn
pandas
